move all existing models to an sql folder to make way for mongoose
