+++ src/models/mysql/discount-type.ts
