+++ src/models/mysql/invoice.ts
