+++ src/models/mysql/paper-type.ts
