+++ src/routes/frames.routes.ts
@@ -1,5 +1,5 @@
 import express from "express";
-import { Frame } from "../models/frame";
+import { Frame } from "../models/mysql/frame";
 import { sendErrorResponse } from "../utils/responses.util";
 
 const router = express.Router();
